requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_core==0.3.72
langchain_ollama==0.3.6
langgraph==0.5.4
pydantic==2.11.7
python-dotenv==1.1.1
